[plugin-prevent-multiple-instances] Change error to warning for yarn run
WHen running yarn workspaces focus this will find a lot of problems are yarn exits error, which we don't want.  Instead let's make this a warning in the normal yarn output and if people want to block this in CI they can run the command to check for conflicts and that'll exit error.